pyproject.toml
[project]
name = "pyconf-hyd-2025"
version = "0.1.0"
description = "Project for Mastering Test Driven Development in Django: A Comprehensive Guide with factory_boy and faker workshop at PyConf Hyderabad 2025"
authors = [
    {name = "Kuldeep Pisda",email = "hello@kdpisda.in"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.1.6,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "django-filter (>=25.1,<26.0)",
    "djangorestframework-simplejwt (>=5.4.0,<6.0.0)",
    "drf-standardized-errors (>=0.14.1,<0.15.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
factory-boy = "^3.3.3"
pytest-django = "^4.10.0"
parameterized = "^0.9.0"
pytest-cov = "^6.0.0"
pytest-xdist = "^3.6.1"
pre-commit = "^4.1.0"
psutil = "^7.0.0"
